requirements.txt
# --- Data Processing ---
pandas==2.1.3
jsonlines==4.0.0
tqdm==4.66.1

# --- Vector Math & Embeddings ---
# Upgrading to 3.0.0+ removes the dependency on the broken legacy code
sentence-transformers==3.0.1
transformers>=4.38.0
torch==2.2.0
numpy==1.26.2
huggingface-hub>=0.23.0

# --- Vector Database ---
faiss-cpu==1.7.4
chromadb==0.4.24

# --- API & Framework ---
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

python-dotenv==1.0.0
google-generativeai>=0.3.0


